backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
google-genai==1.13.0
pydantic-settings==2.9.1
